src/flask_service/requirements.txt
Flask
requests
python-jose[cryptography]
python-dotenv
gunicorn